fix viewport
