+++ src/components/elements/index.ts
@@ -2,3 +2,4 @@ export * from "./CopyToClipboard";
 export * from "./Did";
 export * from "./Icons";
 export * from "./SearchBox";
+export * from "./Viewport";
